Create custom request class (via Volley) that enables caching.
This will allow us to get data from Volley's built-in cache, rather than retry the network for the same data. After 3 minutes the current data will be returned, but the JSON feed will be read again. After 24 hours, the cache is wiped and the JSON feed read.

We also try and get feed data once the fragment view has been created, so that we know we are ready to go when it comes to updating the list.